add code copy function
