+++ src/components/card/sizingToolCard/index.module.less
@@ -99,3 +99,27 @@
     }
   }
 }
+
+.guideWrapper {
+  position: relative;
+  padding-right: 32px;
+  background-color: #f6f6f6;
+  border-radius: 4px;
+
+  .codeWrapper {
+    overflow-x: auto;
+    padding: 0 1em;
+  }
+
+  .copyBtn {
+    position: absolute;
+    right: 8px;
+    top: 8px;
+    width: 14px;
+    height: 14px;
+
+    svg {
+      font-size: 14px;
+    }
+  }
+}
